+++ .eslintrc.js
@@ -5,7 +5,7 @@ var ERROR = 2;
 module.exports = {
 	env: {
 		'node': true,
-		'es6': true
+		'es2017': true
 	},
 
 	extends: 'eslint:recommended',