+++ Alloy/commands/compile/BuildLog.js
@@ -1,5 +1,4 @@
 var fs = require('fs-extra'),
-	chmodr = require('chmodr'),
 	path = require('path'),
 	CONST = require('../../common/constants'),
 	logger = require('../../logger');
@@ -25,7 +24,6 @@ function BuildLog(_projectPath) {
 	// make sure the alloy build folder exists
 	if (!fs.existsSync(dir)) {
 		fs.mkdirpSync(dir);
-		chmodr.sync(dir, 0755);
 	}
 
 	// load it up