+++ Alloy/commands/compile/ast/optimizer-plugin.js
@@ -61,7 +61,7 @@ module.exports = function (_ref) {
 				}
 			},
 			Identifier: function(path) {
-				if (this.defines.hasOwnProperty(path.node.name) &&
+				if (Object.prototype.hasOwnProperty.call(this.defines, path.node.name) &&
 					(path.parent.type !== 'VariableDeclarator' || path.node.name !== path.parent.id.name)) {
 					path.replaceWith(types.booleanLiteral(this.defines[path.node.name]));
 				}