+++ Alloy/commands/compile/compilerUtils.js
@@ -4,14 +4,14 @@ var U = require('../../utils'),
 	os = require('os'),
 	fs = require('fs-extra'),
 	walkSync = require('walk-sync'),
-	chmodr = require('chmodr'),
 	jsonlint = require('jsonlint'),
 	logger = require('../../logger'),
 	astController = require('./ast/controller'),
 	_ = require('lodash'),
 	styler = require('./styler'),
-	XMLSerializer = require('xmldom').XMLSerializer,
-	CONST = require('../../common/constants');
+	XMLSerializer = require('@xmldom/xmldom').XMLSerializer,
+	CONST = require('../../common/constants'),
+	sourceMapper = require('./sourceMapper');
 
 ///////////////////////////////////////
 ////////// private variables //////////
@@ -50,7 +50,7 @@ var RESERVED_ATTRIBUTES = [
 		'method',
 		'module'
 	],
-	RESERVED_EVENT_REGEX =  /^on([A-Z].+)/;
+	RESERVED_EVENT_REGEX =  new RegExp(`^(?:(${CONST.PLATFORMS.join('|')}):)?on([A-Z].+)`);
 
 // load CONDITION_MAP with platforms
 exports.CONDITION_MAP = {
@@ -86,7 +86,7 @@ exports.generateVarName = function(id, name) {
 			'Reserved words: [' + CONST.JS_RESERVED_ALL.sort().join(',') + ']'
 		]);
 	}
-	return '$.__views.' + id;
+	return '$.__views["' + id + '"]';
 };
 
 exports.generateUniqueId = function() {
@@ -206,8 +206,11 @@ exports.getParserArgs = function(node, state, opts) {
 		if (_.includes(attrs, attrName)) { return; }
 		var matches = attrName.match(RESERVED_EVENT_REGEX);
 		if (matches !== null && exports.isNodeForCurrentPlatform(node) && !_.includes(CONST.SPECIAL_PROPERTY_NAMES, attrName)) {
+			if (matches[1] && compilerConfig.alloyConfig.platform !== matches[1]) {
+				return;
+			}
 			events.push({
-				name: U.lcfirst(matches[1]),
+				name: U.lcfirst(matches[2]),
 				value: node.getAttribute(attrName)
 			});
 		} else {
@@ -224,7 +227,7 @@ exports.getParserArgs = function(node, state, opts) {
 				}
 			}	
 
-			if (/(^|\+)\s*(?:(?:Ti|Titanium|Alloy.Globals|Alloy.CFG|\$.args)\.|L\(.+\)\s*$)/.test(theValue)) {
+			if (/(^|\+)\s*(?:(?:Ti|Titanium|Alloy.Globals|Alloy.CFG|\$.args)\.|L\(.+\)\s*$|WPATH\()/.test(theValue)) {
 				var match = theValue.match(/^\s*L\([^'"]+\)\s*$/);
 				if (match !== null) {
 					theValue = theValue.replace(/\(/g, '("').replace(/\)/g, '")');
@@ -632,12 +635,25 @@ exports.copyWidgetResources = function(resources, resourceDir, widgetId, opts) {
 				var dest = path.join(destDir, path.basename(file));
 				if (!path.existsSync(destDir)) {
 					fs.mkdirpSync(destDir);
-					chmodr.sync(destDir, 0755);
 				}
 
 				logger.trace('Copying ' + file.yellow + ' --> ' +
 					path.relative(compilerConfig.dir.project, dest).yellow + '...');
 				U.copyFileSync(source, dest);
+
+				if (path.extname(source) === '.js' && compilerConfig.sourcemap) {
+					sourceMapper.generateSourceMap({
+						target: {
+							filename: file,
+							filepath: dest,
+						},
+						data: {},
+						origFile: {
+							filename: file,
+							filepath: source
+						}
+					}, compilerConfig);
+				}
 			}
 		});
 
@@ -698,8 +714,9 @@ exports.mergeI18N = function mergeI18N(src, dest, opts) {
 			if (!fs.existsSync(srcFile)) return;
 
 			if (fs.statSync(srcFile).isDirectory()) {
-				fs.existsSync(destFile) || fs.mkdirpSync(destFile);
-				chmodr.sync(destFile, 0755);
+				if (!fs.existsSync(destFile)) {
+					fs.mkdirpSync(destFile);
+				}
 				return walk(srcFile, destFile);
 			}
 
@@ -734,7 +751,7 @@ exports.mergeI18N = function mergeI18N(src, dest, opts) {
 
 			_.each(srcDoc.getElementsByTagName('string'), function (node) {
 				var name = node.getAttribute('name');
-				if (!existing.hasOwnProperty(name)) {
+				if (!Object.prototype.hasOwnProperty.call(existing, name)) {
 					destDoc.appendChild(destXml.createTextNode('\t'));
 					destDoc.appendChild(node);
 					destDoc.appendChild(destXml.createTextNode('\n'));
@@ -865,7 +882,6 @@ function generateConfig(obj) {
 		buildLog.data.cfgHash = hash;
 		// write out the config runtime module
 		fs.mkdirpSync(resourcesBase);
-		chmodr.sync(resourcesBase, 0755);
 
 		//logger.debug('Writing "Resources/' + (platform ? platform + '/' : '') + 'alloy/CFG.js"...');
 		var output = 'module.exports=' + JSON.stringify(o) + ';';
@@ -875,7 +891,6 @@ function generateConfig(obj) {
 		var baseFolder = path.join(obj.dir.resources, 'alloy');
 		if (!fs.existsSync(baseFolder)) {
 			fs.mkdirpSync(baseFolder);
-			chmodr.sync(baseFolder, 0755);
 		}
 		fs.writeFileSync(path.join(baseFolder, 'CFG.js'), output);
 	}