+++ Alloy/commands/compile/index.js
@@ -2,7 +2,6 @@ var ejs = require('ejs'),
 	path = require('path'),
 	fs = require('fs-extra'),
 	walkSync = require('walk-sync'),
-	chmodr = require('chmodr'),
 	vm = require('vm'),
 	babel = require('@babel/core'),
 	async = require('async'),
@@ -255,7 +254,6 @@ module.exports = function(args, program) {
 	_.each(['COMPONENT', 'WIDGET', 'RUNTIME_STYLE'], function(type) {
 		var p = path.join(paths.resources, titaniumFolder, 'alloy', CONST.DIR[type]);
 		fs.mkdirpSync(p);
-		chmodr.sync(p, 0755);
 	});
 
 	// Copy in all developer assets, libs, and additional resources
@@ -363,7 +361,6 @@ module.exports = function(args, program) {
 		fs.removeSync(destPlatformDir);
 	}
 	fs.mkdirpSync(destPlatformDir);
-	chmodr.sync(destPlatformDir, 0755);
 	fs.writeFileSync(path.join(destPlatformDir, 'alloy_generated'), generateMessage('platform'));
 	sourcePlatformDirs.forEach(function (dir) {
 		var dirs = [ dir ];
@@ -388,7 +385,6 @@ module.exports = function(args, program) {
 		fs.removeSync(destI18NDir);
 	}
 	fs.mkdirpSync(destI18NDir);
-	chmodr.sync(destI18NDir, 0755);
 	fs.writeFileSync(path.join(destI18NDir, 'alloy_generated'), generateMessage('i18n'));
 	sourceI18NPaths.forEach(function (dir) {
 		if (fs.existsSync(dir)) {
@@ -479,6 +475,8 @@ module.exports = function(args, program) {
 	logger.info('');
 
 	generateAppJs(paths, compileConfig, restrictionPath, compilerMakeFile);
+
+	U.copyFileSync(path.join(alloyRoot, 'template', 'alloy.bootstrap.js'), path.join(paths.resources, titaniumFolder, 'alloy.bootstrap.js'));
 
 	// ALOY-905: workaround TiSDK < 3.2.0 iOS device build bug where it can't reference app.js
 	// in platform-specific folders, so we just copy the platform-specific one to
@@ -908,14 +906,10 @@ function parseAlloyComponent(view, dir, manifest, noView, fileRestriction) {
 			path.join(compileConfig.dir.resources, titaniumFolder, 'alloy', CONST.DIR.WIDGET,
 				manifest.id, widgetDir)
 		);
-		chmodr.sync(path.join(compileConfig.dir.resources, titaniumFolder, 'alloy', CONST.DIR.WIDGET,
-			manifest.id, widgetDir), 0755);
 		fs.mkdirpSync(
 			path.join(compileConfig.dir.resources, titaniumFolder, 'alloy', CONST.DIR.WIDGET,
 				manifest.id, widgetStyleDir)
 		);
-		chmodr.sync(path.join(compileConfig.dir.resources, titaniumFolder, 'alloy', CONST.DIR.WIDGET,
-			manifest.id, widgetStyleDir), 0755);
 
 		// [ALOY-967] merge "i18n" dir in widget folder
 		CU.mergeI18N(path.join(dir, 'i18n'), path.join(compileConfig.dir.project, 'i18n'), { override: false });
@@ -1019,7 +1013,6 @@ function parseAlloyComponent(view, dir, manifest, noView, fileRestriction) {
 		styleCode += _.template(fs.readFileSync(path.join(alloyRoot, 'template', 'wpath.js'), 'utf8'))({ WIDGETID: manifest.id });
 	}
 	fs.mkdirpSync(path.dirname(runtimeStylePath));
-	chmodr.sync(path.dirname(runtimeStylePath), 0755);
 	fs.writeFileSync(runtimeStylePath, styleCode);
 }
 
@@ -1099,7 +1092,6 @@ function processModels(dirs) {
 					titaniumFolder, 'alloy', 'widgets', manifest.id, 'models');
 			}
 			fs.mkdirpSync(modelRuntimeDir);
-			chmodr.sync(modelRuntimeDir, 0755);
 			fs.writeFileSync(path.join(modelRuntimeDir, casedBasename + '.js'), code);
 			models.push(basename);
 		});